requirements.txt
# Core dependencies
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
PyYAML>=6.0
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

